+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* Updated `tbl_regression()` to be compatible with models created with the {parsnip} and {workflows} packages (#890, #647)
+
 * Added the `modify_table_styling(text_format = "indent2")` option to double indent a row. (#864)
 
 * Messaging update when `inline_text.gtsummary()` suspects a variable has been repeated in the gtsummary table. (#855)